remove extra ":" from prefixed groups
Prefixes with a colon, for example the default of `asgi:`, were getting a duplicate colon, `prefix::group:name` instead of `prefix:group:name`. If prefix was defined without a colon, for example, `asgi` then the channels were smashed together, `asgispecific.14124asf!`

I think this would be a breaking change for current users which isn't great, but it would make all the keys consistent and easier to parse if needed.